bug fix in terminal condition

+++ code/cpp/Game.hpp
@@ -390,7 +390,7 @@ class Game {
     // return 1.0 - minDistToGoal /(1.41 * 0.25);
 
     int numAttackerActive = 0;
-    int reachedGoal = 0;
+    float reachedGoal = 0.0;
     for (const auto& attacker : state.attackers) {
       if (   attacker.status == RobotStateT::Status::Active
           || attacker.status == RobotStateT::Status::ReachedGoal) {
